python-pdf-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pdfplumber==0.11.0
PyMuPDF==1.23.26
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
boto3==1.34.0
google-cloud-documentai==2.24.0
google-cloud-storage==2.14.0
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==3.0.1
